+++ unit_tests/storage/data_layout.cpp
@@ -3,10 +3,10 @@
 #include "../common/range_tools.hpp"
 #include "../common/temporary_file.hpp"
 
-#include <boost/filesystem.hpp>
 #include <boost/test/unit_test.hpp>
 
 #include <cmath>
+#include <filesystem>
 
 BOOST_AUTO_TEST_SUITE(data_layout)